fix: reorganize import to avoid circular dependency
